backend/fastapi-text-service/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf  
python-pptx 
python-docx  

